examples/marketplace/src/pyproject.toml
[project]
name = "marketplace-example"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
